Adjust malloc size calculation in philo_init
This commit corrects an issue in philo_init.c where memory was improperly allocated for philosopher objects. The '+' operator was mistakenly used instead of parentheses, causing over-allocation of memory.

Now, the malloc function correctly calculates the size to be 'number of philosophers + 1'. This fix helps to avoid potential issues due to memory overrun and it notably improves the efficiency of the code by preventing unnecessary allocation.