Change variable types from size_t to long and ft_atoi to ft_atol
Changed variable declaration from size_t to long hence ft_atoi function to ft_atol in several files to handle negative numbers. This change was necessary for correct handling of invalid argument situations such as the case of negative numbers, which were not considered as invalid in the old code. Also, modified error messages to be more informative and accurate.