Added gprof-helper.c to profile pthread-enabled programs
A new file, gprof-helper.c, has been added to the source. This file includes a preload library to support profiling of pthread-enabled programs. The aim is to boost debugging efficiency by enabling the user to gather more specific profiling data. In the library, the original pthread was wrapped around a structure that sets the timer value to aid profiling. Use this by preloading with your pthread enabled program. The license allows free use and modification.